xds: move bootstrap config generating utility package to testutils
